proyectoCuestionario/requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.7
django-bootstrap4==0.0.4
django-debug-toolbar==1.9
django-filter==1.1.0
django-guardian==1.4.9
djangorestframework==3.7.3
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.9
MarkupSafe==1.0
MySQL-python==1.2.5
py==1.5.2
pytest==3.2.5
pytest-django==3.1.2
pytz==2017.3
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.22
psycopg2==2.7.3.2